feat(dapps): Redesign discover tab and dapps screen (#5569)
### Description
 
For
[RET-1089](https://linear.app/valora/issue/RET-1089/[wallet]-implement-explore-dapps-section-on-discover-page).
Redesigns the Discover tab and moves the primary dapp section back to
its own screen.

### Test plan

Manual and unit tested. See video below.



https://github.com/valora-inc/wallet/assets/569401/7a7ef68c-d177-4c36-837f-ba901fd6e346


### Related issues

- Fixes
[RET-1089](https://linear.app/valora/issue/RET-1089/[wallet]-implement-explore-dapps-section-on-discover-page).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)