chore: update discover dapp and dapp filter styles (#5625)
### Description


[Context](https://valora-app.slack.com/archives/C0684TXDR3K/p1719510075150019)
and
[design](https://www.figma.com/design/FwRyq9HaKtZgywKVmIg0Zh/Discover-Page?node-id=1-3&t=uHdxyczMjeuYYBO2-0).

This PR only has styles changes. Also fixes the dapp filters appearing
on more than 1 row. This PR breaks the dapp rankings styles but that is
disabled and earmarked for deletion anyway.

### Test plan

![Simulator Screenshot - iPhone 15 Pro - 2024-07-09 at 12 16
46](https://github.com/valora-inc/wallet/assets/20150449/c34c0855-6b06-4384-83e6-b18329f2e63a)


https://github.com/valora-inc/wallet/assets/20150449/7f1cd23b-8e5f-497a-88e9-174ad7afdf5a



### Related issues

n/a

### Backwards compatibility

yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)