feat(earn): update earn discover card for multiple pools (#5649)
### Description

Sets up the statsig gate and shows the earn entrypoint card on the
discover tab. Only changes the card for when there are no active pools.
Also refactors the EarnInfoScreen to not take tokenId as a param because
it is a generic screen.


[Figma](https://www.figma.com/design/E1rC3MG74qEg5V4tvbeUnU/Stablecoin-Enablement?node-id=4079-14115&t=VKwjwdsC2mZD4NcH-0)

### Test plan

<img
src="https://github.com/user-attachments/assets/a99e5980-f5dc-4da6-8e44-f5b191fd4e2d"
width="250" />


### Related issues

- Part of ACT-1257

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)