feat: EstimatedNetworkFee with loading state and multiple TXs (#5683)
### Description

Extend `<EstimatedNetworkFee />` so it can show a loading state (while
preparing transactions) and use multiple TXs to show the total estimated
fee.

It also now uses `getEstimatedFee` instead of `getMaxGasFee`. 
So it's closer to the actual fee.

### Test plan

- Updated tests


https://github.com/user-attachments/assets/e351f231-5885-435f-8555-07df348f9f48

Note: This video is just to show the loading state, this PR doesn't yet
add it to the claim shortcut bottom sheet, it will be in another PR.

### Related issues

- Fixes RET-1108

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)