feat: cross chain swap transaction feed item (#5677)
### Description

This PR augments the swap tx feed item to also display cross chain
swaps.

There is some coordination needed to show the estimated "in" amount for
cross chain swaps, because this information is in the standby
transaction but usually we use the blockchain-api transactions as the
source of truth. I settled on using standby transactions as the source
of truth (if possible) for pending cross chain transactions. we can
eventually decide to use standby transactions as the source of truth for
_all_ pending transactions, but blockchain-api does not have real
examples for this yet.

The above means that:
- when a pending cross chain tx is created from the wallet, we have most
things we need to display the pending tx
- when the pending tx comes back from blockchain-api, it contains the
swap fee info but _not_ the inAmount value. this problem is resolved in
the reducer explicitly by modifying the existing standby transaction.
- when the tx comes back from blockchain-api and is completed, the
standby transaction is removed as usual
- if there is a wallet restore and the standby transaction is removed
before it is settled, the pending tx from blockchain-api goes through
the normal completed tx flow of being rendered using the available info
(no inAmount value)

### Test plan


https://github.com/user-attachments/assets/cb4e571a-1930-4c2a-a374-24537f57ef1f

After the transaction has settled:


https://github.com/user-attachments/assets/9f22457d-aa62-40f6-bab2-3cbc46f7ec35


### Related issues

- Fixes RET-1118

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)