fix: ensure allowedNetworks does not trigger unnecessary rerenders (#5699)
### Description

Rerenders are not (usually) the enemy but this `allowedNetworks` array
is used in a few dependency arrays for computations on the homefeed
which is slow already. I noticed that the `allowedNetworks` array was
always triggering updates (can tell by logging something in a useEffect
or useMemo with only `allowedNetworks` in the dependency array, the logs
are printed frequently) and i think it is because the array does not
preserve its identity when returned from the `getAllowedNetworkIds`
function. I wasn't sure if it's safe to only get the allowed networks on
mount (in case this screen is never remounted and the dynamic config
changes) so opted for a hacky string solution here.

### Test plan

When logging something in a useEffect or useMemo with only
`allowedNetworks` in the dependency array, the log should only be
printed once.

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)