fix(homefeed): prevent reordering on app load (#5732)
### Description

It was annoying me that the homefeed reorders itself on app start. The
root cause is that we initialise the `useFetchTransactions` hook with
empty transactions, then append new transactions to the empty array as
we receive results from each network. This means the first page of
results from the first network response will overwrite all the
transactions that are currently rendered (aside from the standby
transactions, which come through a different datasource). This issue was
exacerbated by the change to show responses per network as we get them.

Reminder of the problem:


https://github.com/user-attachments/assets/0731a96f-23fb-41bd-aed2-cb823575bc56



### Test plan

Not much to show, aside from the tx feed is stable on app start now.

### Related issues

n/a

### Backwards compatibility

n/a

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)