chore: remove hard coded colours (#5709)
### Description

This PR removes all hard coded color values in the app. Most of the
colors had close matches in the `colors` enum already, so nothing much
to call out here.

### Test plan

n/a

### Related issues

- Fixes RET-1155

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)