fix(points): hide live links when user has no jumpstart tokens (#5751)
### Description

Hide live links activity when a user has no jumpstart tokens (or live
links are disabled).
This is consistent with the send flow behavior when the live link option
is not shown when a user has no jumpstart tokens.

If the user tries to create a jumpstart link without having the
jumpstart tokens the app will crash.

Context:
https://valora-app.slack.com/archives/C04B61SJ6DS/p1723075264724759

### Test plan

* Tested manually
* Updated unit test

### Related issues

NA

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)