chore: remove deprecated h1 font (#5734)
### Description

As the title. It seems like a straight replace for h1 -> titleMedium,
this is what was asked for in RET-1165.

<img width="397" alt="Screenshot 2024-08-08 at 15 10 08"
src="https://github.com/user-attachments/assets/0dede232-d83c-41ca-a7b0-3110ac13029f">


### Test plan

n/a

### Related issues

Related to RET-1174 

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)