chore: remove support for old escrow invitations flow (#5778)
### Description

This PR removes support for the old escrow flow, where you could send
funds to the escrow contract to invite a friend (using their phone
number). Specifically, the following is deleted:
- support for reclaiming sent funds. this includes the UI notification
in notification center, as well as all the backend sagas and actions.
- support for showing invitations via escrow in the homefeed. it's been
a long time since we removed escrow so not many users should be impacted
by this change. those transactions will be omitted from the feed.

### Test plan

<!-- Demonstrate the change is solid, or why it doesn't need testing.
Example: add any manual testing steps or scenarios (if not obvious),
screenshots / videos if the pull request changes the user interface.
-->

### Related issues

- Related to RET-1173

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)