refactor: make generic add assets bottom sheet component (#5805)
### Description

As the title. This change makes it easier to implement a similar
component for the points
[design](https://www.figma.com/design/6sJ4fCxCptTybHB6ZZ7Pqi/Escrow-Feature?node-id=3341-3388&t=wNp4q9sJM0JyZxXQ-0).

The new component doesn't contain much logic so I didn't invest time in
making unit tests.

### Test plan

n/a

### Related issues

- Related to RET-1177

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)