refactor: move addAccount and loadStoredAccounts to KeychainLock (#5809)
### Description

Move remaining logic for account management from `KeychainWallet`
(ContractKit specific) to `KeychainLock` (library agnostic).
This will make it easier to replace `KeychainWallet` with `KeychainLock`
once we remove ContractKit.

### Test plan

- Updated tests

### Related issues

- Related to RET-1182

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)