refactor: remove branding folder (#5813)
### Description

This PR:
1. remove branding folder and move all assets to the intended `android`,
`ios`, `src` directories
2. in the `images` folder, put all `.png` files in an `assets` sub
folder. these images continue to be exported from `Images.ts`.
3. move "branded" components from `icons` to `images` - these are all
files that use custom colours, but eventually should be all icons that
are not generic (e.g. Material UI icons)
4. remove branding script and references to it

### Test plan

n/a

### Related issues

- Fixes RET-1156

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Alex Bakoushin <alex@bakoush.in>