chore: consolidate and remove unnecessary icons (#5826)
### Description

After implementing the debug screen, I noticed some duplicate icons.
This PR removes the duplicates and replaces them with similar
alternatives. There should be no visual changes in this PR.

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)