fix: Use 2 letter code for hooks API language header (#5846)
### Description

`hooks` repo `.crowdin.yml`
([here](https://github.com/valora-inc/hooks/blob/main/.crowdin.yml))
uses 2 letter code for language, update API header to use that. Same as
`dapp-list`.

### Test plan

Unit tests updated

### Related issues

- Related to ACT-1331

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)