feat(settings): Implement Profile submenu + redesign, remove profile picture (#5803)
### Description

For
[RET-1180](https://linear.app/valora/issue/RET-1180/[settings]-create-new-settings-screen).
Implements the profile redesign within the new Settings screen, while
removing the profile picture feature. The new settings menu is still not
accessible, however some existing components were modified, resulting in
a few visible changes to the app. The main ones are:

* Existing profile pictures have been removed, as well as the ability to
add new ones.
* The edit profile screen, currently accessible in-app, has been updated
to match the Figma designs.
* Some design changes on the connect phone number screen and revoke
phone number bottom sheet

### Test plan

Unit and manual tested. See videos below (first video is before removing
picture but shows everything, second video just demonstrates removal of
profile picture).



https://github.com/user-attachments/assets/d1c086b9-6503-4686-bba2-5a350e831651



https://github.com/user-attachments/assets/34e83696-b542-40a2-9d8b-7120cd22d510



### Related issues

- Part of
[RET-1180](https://linear.app/valora/issue/RET-1180/[settings]-create-new-settings-screen)

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)