chore: copy used code from @celo/phone-utils into the repo and remove @celo/phone-utils dependency (#5852)
### Description
Copy used functions/types from
[@celo/phone-utils](https://www.npmjs.com/package/@celo/phone-utils) and
also depending functions/types:
- E164Number
- E164PhoneNumberType
- isE164NumberStrict
- Countries
- LocalizedCountry
- getCountryEmoji
- getRegionCode
- getRegionCodeFromCountryCode
- getDisplayNumberInternational
- getPhoneHash from `@celo/phone-utils/lib/getPhoneHash`
- validateInput
- parsePhoneNumber

PhoneNumberUtils was not copied as it is just an object exporting some
functions from the list above and the only function used from it was
`getPhoneHash` so it was changed to use the function directly.

For `getPhoneHash` a legacy implementation of `getIdentificationHash`
dependency was used:

https://github.com/celo-org/developer-tooling/blob/8572a0f978d1aa01a36775ef4be48c3eafdbb204/packages/sdk/base/src/identifier.ts
Latest version of implementation:

https://github.com/celo-org/social-connect/blob/main/packages/odis-identifiers/src/identifier.ts
P.S. The only difference between the two is `IdentifierPrefix` having an
extra `GITHUB` enum value in the latest version.

#### New dependencies:
-
[google-libphonenumber](https://github.com/ruimarinho/google-libphonenumber)
- [io-ts](https://github.com/gcanti/io-ts)
- [fp-ts](https://github.com/gcanti/fp-ts)

### Test plan
All tests pass with the new copied functions.

### Related issues

- Relates to RET-1186

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)