refactor(aave-earn-collect): Use hooks to gather pool information for aave collect & withdraw screen (#5856)
### Description

This is part of an effort to refactor the earn withdraw flow so that it
can be used for multipool. This involves both making the flow generic
and not aave-specific and utilizing hooks. I've broken the work down
into three steps that I'll take. This Pr encompasses step 1

1. Refactor the "read" part of the withdraw flow (reading what rewards
are available, apy, etc.) to come from hooks data
2. Refactor the "write" part of the withdraw flow to come from hooks
3. Refactor the withdraw flow to be not aave-specific and hook it up to
the multipool flow.

### Test plan

Test along each step of the way if the flow works for both the
aave-specific single pool flow as well as the multipool flow

### Related issues

https://linear.app/valora/issue/ACT-1343/refactor-earn-withdraw-flow