feat(settings): Add in Security submenu to new settings screen (#5839)
### Description

For
[RET-1180](https://linear.app/valora/issue/RET-1180/%5Bsettings%5D-create-new-settings-screen).
Adds in the Security & Privacy submenu in the new settings screen.
Includes the old dev menu as normal by tapping the submenu title five
times. Various small styling tweaks to a handful of flows that are
accessible from the Security & Privacy screen.

Do not be alarmed by the diff size! The new `SecuritySubmenu.tsx` and
associated test file are both basically entirely ported over from
`Settings.tsx`. The next PR in this (very long) chain will hook up the
new Settings experience to the rest of the app, and remove our old one
entirely, resulting in a lot of code gone.

### Test plan

Unit and manual tested. See video below.


https://github.com/user-attachments/assets/ea50b1da-3fcd-490c-ac05-6197c8c8515f


### Related issues

- Part of
[RET-1180](https://linear.app/valora/issue/RET-1180/%5Bsettings%5D-create-new-settings-screen).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)