refactor: update and simplify account util with latest version from @celo/cryptographic-utils (#5863)
### Description

This pulls the latest changes from the
[account](https://github.com/celo-org/developer-tooling/blob/467d4e16444535d341bd2296d41c386f1dab187f/packages/sdk/cryptographic-utils/src/account.ts)
module from `@celo/cryptographic-utils`.

Updated slightly to remove unused exports / functions and use viem's
helpers when possible.

### Test plan

- Tests pass

### Related issues

- Part of RET-1185

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)