chore(deps): remove @ethereumjs/utils (#5912)
### Description
Replace left-over usage of `@ethereumjs/utils` with existing equivalent
functions or with functions from `viem`

### Test plan
The only affected function is `uriDataFromJson` from
`src/qrcode/schema.ts`.
All tests of this function from `src/qrcode/schema.test.ts` pass.

### Related issues
N/A

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)