feat(earn): Multi-pool earn analytics once over (#5886)
### Description

Update to make properties consistent, check that there is test coverage
for all events.

Note: No test coverage for `earn_enter_amount_info_more_pools`, but that
is for the single pool flow only.

### Test plan

Unit tests updated

### Related issues

- Fixes ACT-1351

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)