fix: use mocked date in EarnPoolInfoScreen test (#5942)
### Description

As the title.

Also fixes the following warning in the tests by updating the way we are
using `MockedNavigator` in the tests.
```
Looks like you're passing an inline function for 'component' prop for the screen 'MockedScreen' (e.g. component={() => <SomeComponent />}). Passing an inline function will cause the component state to be lost on re-render and cause perf issues since it's re-created every render. You can pass the function as children to 'Screen' instead to achieve the desired behaviour.
```

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)