fix(e2e): Fix funding script (#5976)
### Description

In an effort to [remove Celo dependencies from our e2e
tests](https://github.com/valora-inc/wallet/pull/5782), the way we check
wallet balances was changed in a subtle way such that wallet balances
were being returned as a map from token address to balance rather than
token symbol to balance. This caused our funding script to silently stop
working, despite never actually failing (it would just decide that the
e2e wallets don't need any tokens, and succeed without attempting to
send funds). This PR restores the original behavior, fixing the funding
scripts.

### Test plan

Manual tested

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)