feat(navigation): Update navigation (#10)
### Description

For
[RET-1148](https://linear.app/valora/issue/RET-1148/[shell]-update-navigation-homewalletactivity).
Updates navigation and home tabs to match designs. Also removes the
Valora gradient from certain underlines throughout the app.

### Test plan

Unit and manual tested. See video below.



https://github.com/user-attachments/assets/c9c3e735-665c-4454-abe9-5ce0e4e06500



### Related issues

Fixes
[RET-1148](https://linear.app/valora/issue/RET-1148/[shell]-update-navigation-homewalletactivity).

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)