refactor: reuse same variable for user agent header and app registry (#6017)
### Description

As the title. This removes the hard coded "celo" value for the app names
in the native code.

### Test plan

CI passes

### Related issues

- Related to RET-1157

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)