fix: iOS release workflow (#6028)
### Description


[Context](https://valora-app.slack.com/archives/C029Z1QMD7B/p1726668187770699?thread_ts=1726666233.603359&cid=C029Z1QMD7B).
The `PRODUCT_NAME` of the iOS app is now being set via the
`APP_REGISTRY_NAME` config variable, but the CI was not able to get this
value and the `.ipa` file generated did not have a name. The empty
string name caused the `upload_to_testflight` step of the release CI to
fail.

This PR adds the expected config variable to the FastFile and sets the
`.ipa` file name explicitly.

### Test plan

[Tested the these changes work for
Alfajores](https://github.com/valora-inc/wallet/actions/runs/10924192572/job/30322633647).

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Jean Regisser <jean.regisser@gmail.com>