feat(earn): Add earn_supported_pools dynamic config (#6112)
### Description

and use in getEarnPositions query. Opted to have the default be the
empty list and have the legacy default live in hooks
([here](https://github.com/valora-inc/hooks/pull/598)). That way if we
need to change the legacy default we can and there won't be BC issues in
the wallet (can just change on hooks side without release, goes to all
versions).

Dynamic config:
https://console.statsig.com/4plizaPmWwPL21ASV4QAO0/dynamic_configs/supported_earn_pools

### Test plan

CI

### Related issues

- Fixes ACT-1367

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)