feat(feed): Add initial implementation of Transaction Feed V2 (#6135)
### Description
2nd PR for RET-1207. Implements the following list of basic Transactions
Feed functionality:
- Adds fetching of all the pages for the wallet address up to the point
when there are no transaction to fetch anymore (not showing "no
transactions" toast yet, will be added in the follow-up PR).
- Cursor for the next page is the timestamp of the last transaction from
the current page. If next page includes the same transaction - it gets
deduplicated.
- Re-uses pending transactions from `pendingStandByTransactionsSelector`
- Polls first page every 10 seconds
- Sorts transactions to show approvals at the top if the corresponding
transaction has identical timestamp

More comments for different parts of the pagination flow are added as
comments in the file in the [next
PR](https://github.com/valora-inc/wallet/pull/6136).

### Test plan
10 out of 16 tests from `TransactionFeed.test.ts` were re-used and
adjusted to the new data fetching flow. Other tests will be added in the
follow-up PRs purely for the sake of trying to keep the PRs smaller.

### Related issues

- Relates to RET-1207

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)