feat(feed): Add RTK-Query, add api slice for Transaction Feed V2 (#6134)
### Description
1st PR for RET-1207. It adds RTK-Query and api slice for the new
Transaction Feed V2.

All the reducers up to this point were always manually
maintained/changed by us. For this purpose, we have Redux migrations
which are covered with tests to ensure compatibility between different
versions of Redux state. With introduction of RTK-Query we introduce
library-managed api reducers, which _**must not**_ be manually tweaked
as library can change how it stores its structure (api reducer is a
**cache** for API endpoint).

For this purpose, it is mandatory to omit api reducers completely from
the flow of `typescript-json-schema` generation. This implemenation
moves existing reducers to a separate file as generating schema with api
reducers throws an error of unsupported type from
`@reduxjs/tookit/query/react`.

This new Transaction Feed V2 (that is gonna be implemented over a few
PRs) is intended to completely replace existing
[TransactionFeed](https://github.com/valora-inc/wallet/blob/main/src/transactions/feed/TransactionFeed.tsx),
[queryHelper](https://github.com/valora-inc/wallet/blob/main/src/transactions/feed/queryHelper.ts)
and corresponding test files.
It will be closed behind the feature gate from Statsig for the initial
release.

### Test plan
With this implementation, all the existing tests are working and there
is no issues with persistence as the api reducers are completely omitted
from the persistence flow. Transactions persistence for Transaction Feed
V2 will be implemented in the follow-up PRs.

### Related issues

- Relates to RET-1207

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)