chore: add script to generate release notes (#6131)
### Description

Here's an improved version of a basic script I was using locally to
generate release notes.

Features:
- generate release notes for both "main" and "patch" releases
- excludes cherry-picked commits from the previous release
- detects tag prefix so it also works for Mobile Stack (i.e. no
hardcoded "valora")
- verbose mode by default showing what it does (see screenshot), printed
to `stderr` so we can pipe `stdout` and get just the release notes

Example usages:
- `yarn generate-release-notes` this is the recommended usage
- `yarn generate-release-notes --lastTag valora-v1.94.2`
- `yarn generate-release-notes --lastTag valora-v1.94.2 --toRef someRef`

In the future, the plan is to use it from our release workflow to create
the GitHub Release automatically.

We could also update the nightly workflow to use it.

### Test plan

Successfully used it to generate the release notes for
[1.95.0](https://github.com/valora-inc/wallet/releases/tag/valora-v1.95.0)

<img width="679" alt="Screenshot 2024-10-04 at 16 25 16"
src="https://github.com/user-attachments/assets/c55a10d8-d5b0-4f06-81d8-575ae3677e80">

Successfully used it to generate the release for 1.95.1

<img width="633" alt="Screenshot 2024-10-04 at 17 40 19"
src="https://github.com/user-attachments/assets/e155460f-cb1b-4a64-8bb2-5e2f6b97e4a4">


### Related issues

- Slack
[thread](https://valora-app.slack.com/archives/C025V1D6F3J/p1724326960482299)
for context.

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)