feat: add feature gate for zerion transaction feed (#6181)
### Description

As the title.

While updating the TabHome tests, I noticed every test was throwing an
act warning (below) so this PR also updates the tests to remove the
errors.
```
Warning: An update to TransactionFeed inside a test was not wrapped in act(...).
    
    When testing, code that causes React state updates should be wrapped into act(...):
    
    act(() => {
      /* fire events that update state */
    });
    /* assert on the output */
```

### Test plan

Tested via unit tests and manually.

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)