chore(TransactionFeedV2): Connect to new API (#6182)
### Description
Last PR for RET-1207. Connects TransactionFeedV2 to new API with the
corresponding changes to the fetching process.

### Test plan
Fixed all existing tests to pass with the new cursor structure. One test
related to haptic feedback is disabled for now as some flaws in the
logic were revealed. It will be addresses as a separate change and it is
not urgently necessary for the MVP.

### Related issues

- Relates to RET-1207

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)