refactor: remove usage of `__typename` for the tx feed (#6186)
### Description

`__typename` was inherited from GraphQL and is not needed anymore.
It was duplicating the `type` property and making it a bit difficult to
see which one should be used now.

### Test plan

- Tests pass

### Related issues

N/A

### Backwards compatibility

Almost, but the analytics event `earn_feed_item_select` will now return
the `type` value instead of `__typename` for the `origin` property.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)